The build has been deployed into folders.
