+++ backend/app/Http/Resources/Admin/ClassRoom/ClassRoomResource.php
@@ -1,23 +1,22 @@
 <?php
 
-namespace App\Http\Resources\Admin;
+namespace App\Http\Resources\Admin\ClassRoom;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class LanguageResource extends JsonResource
+class ClassRoomResource extends JsonResource
 {
     /**
-     * Transform the resource collection into an array.
+     * Transform the resource into an array.
      *
      * @param  \Illuminate\Http\Request  $request
+     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
      */
     public function toArray($request)
     {
         return [
             'id' => $this->id,
             'name' => $this->name,
-            'code' => $this->code,
-            'is_active' => $this->is_active,
         ];
     }
 }
