+++ backend/app/Http/Resources/Admin/Condition/ConditionCategoryResource.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Resources\Admin\Condition;
 
-use App\Http\Resources\Admin\LanguageResource;
+use App\Http\Resources\Admin\Language\LanguageResource;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class ConditionCategoryResource extends JsonResource
