+++ backend/app/Http/Resources/Admin/Language/LanguageResource.php
@@ -1,22 +1,23 @@
 <?php
 
-namespace App\Http\Resources\Admin;
+namespace App\Http\Resources\Admin\Language;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class ClassRoomResource extends JsonResource
+class LanguageResource extends JsonResource
 {
     /**
-     * Transform the resource into an array.
+     * Transform the resource collection into an array.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
      */
     public function toArray($request)
     {
         return [
             'id' => $this->id,
             'name' => $this->name,
+            'code' => $this->code,
+            'is_active' => $this->is_active,
         ];
     }
 }
