+++ backend/app/Http/Resources/Admin/Lesson/LessonResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Resources\Admin;
+namespace App\Http\Resources\Admin\Lesson;
 
 use Illuminate\Http\Resources\Json\JsonResource;
