+++ backend/app/Services/Admin/Language/LanguageService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Admin;
+namespace App\Services\Admin\Language;
 
 use App\Models\Language;
 use App\Services\Base\BaseService;
