+++ backend/app/Services/Admin/Lesson/LessonService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Admin;
+namespace App\Services\Admin\Lesson;
 
 use App\Models\Lesson;
 use App\Services\Base\BaseService;
