+++ backend/tests/Feature/Admin/Lesson/LessonControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Admin;
+namespace Tests\Feature\Admin\Lesson;
 
 use App\Models\Lesson;
 use App\Models\QuestionCategory;
